Address review: move MPS probe helpers, reword the accelerator warning
Move `_mps_supports` and `_needs_cpu_detour` out of `_negative_binomial.py`
into `distributions/_utils.py`, which is the existing utility module for this
package (the review asked for `distributions/utils.py`; the module is private
here, alongside `_constraints.py`).

`_needs_cpu_detour` now resolves `_mps_supports` in the `_utils` namespace, so
the parametrised detour test patches `dist_utils` rather than `nb_module`. The
patch on `nb_module._needs_cpu_detour` stays as it was, since the name is
imported into that module.

Also apply the wording from #3980 to the `mps` warning in `parse_device_args`.
The "less accurate" half is dropped: measured against a float64 reference,
MPS float32 is within float32 noise of CPU float32 on `lgamma`, `digamma`,
`log`, `exp`, `softmax`, `sum`, `matmul` and on a negative-binomial ELBO
gradient. Operator coverage is a real limitation and is still warned about,
now together with the fact that RNG streams differ per backend.
