+++ src/scvi/distributions/_negative_binomial.py
@@ -1,8 +1,6 @@
 from __future__ import annotations
 
 import warnings
-from functools import cache
-from typing import TYPE_CHECKING
 
 import torch
 import torch.nn.functional as F
@@ -18,9 +16,7 @@ from torch.distributions.utils import (
 from scvi import settings
 
 from ._constraints import optional_constraint
-
-if TYPE_CHECKING:
-    from collections.abc import Callable
+from ._utils import _needs_cpu_detour
 
 
 def torch_lgamma_mps(x: torch.Tensor) -> torch.Tensor:
@@ -36,30 +32,6 @@ def torch_lgamma_mps(x: torch.Tensor) -> torch.Tensor:
     lgamma tensor that performs on a copied version of the tensor
     """
     return torch.lgamma(x.contiguous())
-
-
-@cache
-def _mps_supports(op: Callable[[torch.Tensor], torch.Tensor]) -> bool:
-    """Whether ``op`` has an MPS kernel in the torch build in use.
-
-    ``aten::_standard_gamma`` gained one in torch 2.12 and ``aten::poisson`` in
-    2.14, so the answer depends on the installed version. It is probed rather
-    than version-compared because source and nightly builds report versions
-    that a comparison reads wrongly.
-
-    Callers must check ``torch.backends.mps.is_available()`` first; this is only
-    reached from a distribution whose parameters already live on ``mps``.
-    """
-    try:
-        op(torch.ones(1, device="mps"))
-    except (NotImplementedError, RuntimeError):
-        return False
-    return True
-
-
-def _needs_cpu_detour(on_mps: bool, op: Callable[[torch.Tensor], torch.Tensor]) -> bool:
-    """Whether sampling ``op`` has to be routed through the CPU."""
-    return on_mps and not _mps_supports(op)
 
 
 def log_zinb_positive(
